chore(issue-list): GA issue list refactor
