+++ static/app/views/issueList/overview.tsx
@@ -1,4 +1,4 @@
-import {Component} from 'react';
+import {useCallback, useEffect, useMemo, useRef, useState} from 'react';
 import styled from '@emotion/styled';
 import * as Sentry from '@sentry/react';
 import type {Location} from 'history';
@@ -12,10 +12,8 @@ import * as qs from 'query-string';
 
 import {addMessage} from 'sentry/actionCreators/indicator';
 import {fetchOrgMembers, indexMembersByProject} from 'sentry/actionCreators/members';
-import type {Client} from 'sentry/api';
 import ErrorBoundary from 'sentry/components/errorBoundary';
 import * as Layout from 'sentry/components/layouts/thirds';
-import LoadingIndicator from 'sentry/components/loadingIndicator';
 import {extractSelectionParameters} from 'sentry/components/organizations/pageFilters/utils';
 import type {CursorHandler} from 'sentry/components/pagination';
 import QueryCount from 'sentry/components/queryCount';
@@ -24,15 +22,14 @@ import {t, tct} from 'sentry/locale';
 import GroupStore from 'sentry/stores/groupStore';
 import IssueListCacheStore from 'sentry/stores/IssueListCacheStore';
 import SelectedGroupStore from 'sentry/stores/selectedGroupStore';
+import {useLegacyStore} from 'sentry/stores/useLegacyStore';
 import {space} from 'sentry/styles/space';
 import type {PageFilters} from 'sentry/types/core';
 import type {BaseGroup, Group, PriorityLevel, SavedSearch} from 'sentry/types/group';
 import {GroupStatus, IssueCategory} from 'sentry/types/group';
 import type {RouteComponentProps} from 'sentry/types/legacyReactRouter';
-import type {Organization} from 'sentry/types/organization';
 import {defined} from 'sentry/utils';
 import {trackAnalytics} from 'sentry/utils/analytics';
-import {browserHistory} from 'sentry/utils/browserHistory';
 import CursorPoller from 'sentry/utils/cursorPoller';
 import {getUtcDateString} from 'sentry/utils/dates';
 import getCurrentSentryReactRootSpan from 'sentry/utils/getCurrentSentryReactRootSpan';
@@ -40,20 +37,25 @@ import parseApiError from 'sentry/utils/parseApiError';
 import parseLinkHeader from 'sentry/utils/parseLinkHeader';
 import {makeIssuesINPObserver} from 'sentry/utils/performanceForSentry';
 import {decodeScalar} from 'sentry/utils/queryString';
-import type {WithRouteAnalyticsProps} from 'sentry/utils/routeAnalytics/withRouteAnalytics';
-import withRouteAnalytics from 'sentry/utils/routeAnalytics/withRouteAnalytics';
+import useDisableRouteAnalytics from 'sentry/utils/routeAnalytics/useDisableRouteAnalytics';
+import useRouteAnalyticsParams from 'sentry/utils/routeAnalytics/useRouteAnalyticsParams';
 import normalizeUrl from 'sentry/utils/url/normalizeUrl';
-import withApi from 'sentry/utils/withApi';
-import withOrganization from 'sentry/utils/withOrganization';
-import withPageFilters from 'sentry/utils/withPageFilters';
-import withSavedSearches from 'sentry/utils/withSavedSearches';
+import useApi from 'sentry/utils/useApi';
+import {useLocation} from 'sentry/utils/useLocation';
+import {useNavigate} from 'sentry/utils/useNavigate';
+import useOrganization from 'sentry/utils/useOrganization';
+import usePageFilters from 'sentry/utils/usePageFilters';
+import {useParams} from 'sentry/utils/useParams';
+import usePrevious from 'sentry/utils/usePrevious';
 import IssueListTable from 'sentry/views/issueList/issueListTable';
 import {IssuesDataConsentBanner} from 'sentry/views/issueList/issuesDataConsentBanner';
 import IssueViewsIssueListHeader from 'sentry/views/issueList/issueViewsHeader';
+import {useFetchSavedSearchesForOrg} from 'sentry/views/issueList/queries/useFetchSavedSearchesForOrg';
 import SavedIssueSearches from 'sentry/views/issueList/savedIssueSearches';
 import type {IssueUpdateData} from 'sentry/views/issueList/types';
 import {NewTabContextProvider} from 'sentry/views/issueList/utils/newTabContext';
 import {parseIssuePrioritySearch} from 'sentry/views/issueList/utils/parseIssuePrioritySearch';
+import {useSelectedSavedSearch} from 'sentry/views/issueList/utils/useSelectedSavedSearch';
 
 import IssueListFilters from './filters';
 import IssueListHeader from './header';
@@ -76,43 +78,7 @@ const DEFAULT_GRAPH_STATS_PERIOD = '24h';
 const DYNAMIC_COUNTS_STATS_PERIODS = new Set(['14d', '24h', 'auto']);
 const MAX_ISSUES_COUNT = 100;
 
-type Params = {
-  orgId: string;
-};
-
-type Props = {
-  api: Client;
-  location: Location;
-  organization: Organization;
-  params: Params;
-  savedSearch: SavedSearch;
-  savedSearchLoading: boolean;
-  savedSearches: SavedSearch[];
-  selectedSearchId: string;
-  selection: PageFilters;
-} & RouteComponentProps<{}, {searchId?: string}> &
-  WithRouteAnalyticsProps;
-
-type State = {
-  error: string | null;
-  groupIds: string[];
-  issuesLoading: boolean;
-  memberList: ReturnType<typeof indexMembersByProject>;
-  pageLinks: string;
-  /**
-   * Current query total
-   */
-  queryCount: number;
-  /**
-   * Counts for each inbox tab
-   */
-  queryCounts: QueryCounts;
-  queryMaxCount: number;
-  realtimeActive: boolean;
-  selectAllActive: boolean;
-  // Will be set to true if there is valid session data from issue-stats api call
-  query?: string;
-};
+type Props = RouteComponentProps<{}, {searchId?: string}>;
 
 interface EndpointParams extends Partial<PageFilters['datetime']> {
   environment: string[];
@@ -135,254 +101,161 @@ type StatEndpointParams = Omit<EndpointParams, 'cursor' | 'page'> & {
   expand?: string | string[];
 };
 
-class IssueListOverview extends Component<Props, State> {
-  state: State = this.getInitialState();
+function useIssuesINPObserver() {
+  const _performanceObserver = useRef<PerformanceObserver | undefined>(undefined);
 
-  getInitialState() {
-    const realtimeActiveCookie = Cookies.get('realtimeActive');
-    const realtimeActive =
-      typeof realtimeActiveCookie === 'undefined'
-        ? false
-        : realtimeActiveCookie === 'true';
+  useEffect(() => {
+    _performanceObserver.current = makeIssuesINPObserver();
 
-    return {
-      groupIds: [],
-      actionTaken: false,
-      selectAllActive: false,
-      realtimeActive,
-      pageLinks: '',
-      queryCount: 0,
-      queryCounts: {},
-      queryMaxCount: 0,
-      error: null,
-      issuesLoading: true,
-      memberList: {},
+    return () => {
+      if (_performanceObserver.current) {
+        _performanceObserver.current.disconnect();
+      }
     };
+  }, []);
+}
+
+function useSavedSearches() {
+  const organization = useOrganization();
+  const {data: savedSearches = [], isPending} = useFetchSavedSearchesForOrg(
+    {
+      orgSlug: organization.slug,
+    },
+    {enabled: !organization.features.includes('issue-stream-custom-views')}
+  );
+
+  const params = useParams();
+  const selectedSavedSearch = useSelectedSavedSearch();
+
+  return {
+    savedSearches,
+    savedSearchLoading:
+      !organization.features.includes('issue-stream-custom-views') && isPending,
+    savedSearch: selectedSavedSearch,
+    selectedSearchId: params.searchId ?? null,
+  };
+}
+
+const parsePageQueryParam = (location: Location, defaultPage: number = 0) => {
+  const page = location.query.page;
+  const pageInt = Array.isArray(page)
+    ? parseInt(page[0] ?? '', 10)
+    : parseInt(page ?? '', 10);
+
+  if (isNaN(pageInt)) {
+    return defaultPage;
   }
 
-  componentDidMount() {
-    this._performanceObserver = makeIssuesINPObserver();
-    this._poller = new CursorPoller({
-      linkPreviousHref: parseLinkHeader(this.state.pageLinks)?.previous?.href!,
-      success: this.onRealtimePoll,
-    });
+  return pageInt;
+};
 
-    // Wait for saved searches to load so if the user is on a saved search
-    // or they have a pinned search we load the correct data the first time.
-    // But if searches are already there, we can go right to fetching issues
-    if (
-      !this.props.savedSearchLoading ||
-      this.props.organization.features.includes('issue-stream-performance')
-    ) {
-      const loadedFromCache = this.loadFromCache();
-      if (!loadedFromCache) {
-        // It's possible the projects query parameter is not yet ready and this
-        // request will be repeated in componentDidUpdate
-        this.fetchData();
+function IssueListOverview({router}: Props) {
+  const location = useLocation();
+  const organization = useOrganization();
+  const navigate = useNavigate();
+  const {selection} = usePageFilters();
+  const api = useApi();
+  const realtimeActiveCookie = Cookies.get('realtimeActive');
+  const [realtimeActive, setRealtimeActive] = useState(
+    typeof realtimeActiveCookie === 'undefined' ? false : realtimeActiveCookie === 'true'
+  );
+  const [groupIds, setGroupIds] = useState<string[]>([]);
+  const [pageLinks, setPageLinks] = useState('');
+  const [queryCount, setQueryCount] = useState(0);
+  const [queryCounts, setQueryCounts] = useState<QueryCounts>({});
+  const [queryMaxCount, setQueryMaxCount] = useState(0);
+  const [error, setError] = useState<string | null>(null);
+  const [issuesLoading, setIssuesLoading] = useState(true);
+  const [memberList, setMemberList] = useState<ReturnType<typeof indexMembersByProject>>(
+    {}
+  );
+  const undoRef = useRef(false);
+  const pollerRef = useRef<CursorPoller | undefined>(undefined);
+  const actionTakenRef = useRef(false);
+
+  const {savedSearch, savedSearchLoading, savedSearches, selectedSearchId} =
+    useSavedSearches();
+
+  const groups = useLegacyStore(GroupStore);
+  useEffect(() => {
+    const storeGroupIds = groups.map(group => group.id).slice(0, MAX_ISSUES_COUNT);
+    if (!isEqual(groupIds, storeGroupIds)) {
+      setGroupIds(storeGroupIds);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [groups]);
+
+  useIssuesINPObserver();
+
+  const onRealtimePoll = useCallback(
+    (data: any, {queryCount: newQueryCount}: {queryCount: number}) => {
+      // Note: We do not update state with cursors from polling,
+      // `CursorPoller` updates itself with new cursors
+      GroupStore.addToFront(data);
+      setQueryCount(newQueryCount);
+    },
+    []
+  );
+
+  useEffect(() => {
+    pollerRef.current = new CursorPoller({
+      linkPreviousHref: parseLinkHeader(pageLinks)?.previous?.href!,
+      success: onRealtimePoll,
+    });
+  }, [onRealtimePoll, pageLinks]);
+
+  const getQueryFromSavedSearchOrLocation = useCallback(
+    (props: {location: Location; savedSearch: SavedSearch | null}): string => {
+      if (
+        !organization.features.includes('issue-stream-custom-views') &&
+        props.savedSearch
+      ) {
+        return props.savedSearch.query;
       }
-    }
-    this.fetchMemberList();
-    this.props.setRouteAnalyticsParams?.({
-      issue_views_enabled: this.props.organization.features.includes(
-        'issue-stream-custom-views'
-      ),
-    });
-    // let custom analytics take control
-    this.props.setDisableRouteAnalytics?.();
-  }
 
-  componentDidUpdate(prevProps: Props, prevState: State) {
-    if (prevState.realtimeActive !== this.state.realtimeActive) {
-      // User toggled realtime button
-      if (this.state.realtimeActive) {
-        this.resumePolling();
-      } else {
-        this._poller.disable();
+      const {query} = props.location.query;
+
+      if (query !== undefined) {
+        return decodeScalar(query, '');
       }
-    }
 
-    // If the project selection has changed reload the member list and tag keys
-    // allowing autocomplete and tag sidebar to be more accurate.
-    if (!isEqual(prevProps.selection.projects, this.props.selection.projects)) {
-      this.loadFromCache();
-      this.fetchMemberList();
-    }
+      return DEFAULT_QUERY;
+    },
+    [organization.features]
+  );
 
-    const selectionChanged = !isEqual(prevProps.selection, this.props.selection);
-
-    // Wait for saved searches to load before we attempt to fetch stream data
-    // Selection changing could indicate that the projects query parameter has populated
-    // and we should refetch data.
-    if (this.props.savedSearchLoading && !selectionChanged) {
-      return;
-    }
-
-    if (
-      prevProps.savedSearchLoading &&
-      !this.props.savedSearchLoading &&
-      this.props.organization.features.includes('issue-stream-performance')
-    ) {
-      return;
-    }
-
-    if (
-      prevProps.savedSearchLoading &&
-      !this.props.organization.features.includes('issue-stream-performance')
-    ) {
-      const loadedFromCache = this.loadFromCache();
-      if (!loadedFromCache) {
-        this.fetchData();
+  const getSortFromSavedSearchOrLocation = useCallback(
+    (props: {location: Location; savedSearch: SavedSearch | null}): string => {
+      if (!props.location.query.sort && props.savedSearch?.id) {
+        return props.savedSearch.sort;
       }
-      return;
-    }
 
-    const prevUrlQuery = prevProps.location.query;
-    const newUrlQuery = this.props.location.query;
-
-    const prevQuery = this.getQueryFromSavedSearchOrLocation({
-      savedSearch: prevProps.savedSearch,
-      location: prevProps.location,
-    });
-    const newQuery = this.getQuery();
-
-    const prevSort = this.getSortFromSavedSearchOrLocation({
-      savedSearch: prevProps.savedSearch,
-      location: prevProps.location,
-    });
-    const newSort = this.getSort();
-
-    // If any important url parameter changed or saved search changed
-    // reload data.
-    if (
-      selectionChanged ||
-      prevUrlQuery.cursor !== newUrlQuery.cursor ||
-      prevUrlQuery.statsPeriod !== newUrlQuery.statsPeriod ||
-      prevUrlQuery.groupStatsPeriod !== newUrlQuery.groupStatsPeriod ||
-      prevQuery !== newQuery ||
-      prevSort !== newSort
-    ) {
-      this.fetchData(selectionChanged);
-    } else if (
-      !this._lastRequest &&
-      prevState.issuesLoading === false &&
-      this.state.issuesLoading
-    ) {
-      // Reload if we issues are loading or their loading state changed.
-      // This can happen when transitionTo is called
-      this.fetchData();
-    }
-  }
-
-  componentWillUnmount() {
-    const groups = GroupStore.getState() as Group[];
-    if (groups.length > 0 && !this.state.issuesLoading && !this.state.realtimeActive) {
-      IssueListCacheStore.save(this.getCacheEndpointParams(), {
-        groups,
-        queryCount: this.state.queryCount,
-        queryMaxCount: this.state.queryMaxCount,
-        pageLinks: this.state.pageLinks,
-      });
-    }
-
-    if (this._performanceObserver) {
-      this._performanceObserver.disconnect();
-    }
-    this._poller.disable();
-    SelectedGroupStore.reset();
-    GroupStore.reset();
-    this.props.api.clear();
-    this.listener?.();
-  }
-
-  private _performanceObserver: PerformanceObserver | undefined;
-  private _poller: any;
-  private _lastRequest: any;
-  private _lastStatsRequest: any;
-  private _lastFetchCountsRequest: any;
-  private actionTaken = false;
-  private undo = false;
-
-  getQueryFromSavedSearchOrLocation({
-    savedSearch,
-    location,
-  }: Pick<Props, 'savedSearch' | 'location'>): string {
-    if (
-      !this.props.organization.features.includes('issue-stream-custom-views') &&
-      savedSearch
-    ) {
-      return savedSearch.query;
-    }
-
-    const {query} = location.query;
-
-    if (query !== undefined) {
-      return decodeScalar(query, '');
-    }
-
-    return DEFAULT_QUERY;
-  }
-
-  getSortFromSavedSearchOrLocation({
-    savedSearch,
-    location,
-  }: Pick<Props, 'savedSearch' | 'location'>): string {
-    if (!location.query.sort && savedSearch?.id) {
-      return savedSearch.sort;
-    }
-
-    if (location.query.sort) {
-      return location.query.sort as string;
-    }
-    return DEFAULT_ISSUE_STREAM_SORT;
-  }
-
-  /**
-   * Load the previous
-   * @returns Returns true if the data was loaded from cache
-   */
-  loadFromCache(): boolean {
-    const cache = IssueListCacheStore.getFromCache(this.getCacheEndpointParams());
-    if (!cache) {
-      return false;
-    }
-
-    this.setState(
-      {
-        issuesLoading: false,
-        queryCount: cache.queryCount,
-        queryMaxCount: cache.queryMaxCount,
-        pageLinks: cache.pageLinks,
-      },
-      () => {
-        // Handle this in the next tick to avoid being overwritten by GroupStore.reset
-        // Group details clears the GroupStore at the same time this component mounts
-        GroupStore.add(cache.groups);
-        // Clear cache after loading
-        IssueListCacheStore.reset();
+      if (props.location.query.sort) {
+        return props.location.query.sort as string;
       }
-    );
+      return DEFAULT_ISSUE_STREAM_SORT;
+    },
+    []
+  );
 
-    return true;
-  }
-
-  getQuery(): string {
-    return this.getQueryFromSavedSearchOrLocation({
-      savedSearch: this.props.savedSearch,
-      location: this.props.location,
+  const query = useMemo((): string => {
+    return getQueryFromSavedSearchOrLocation({
+      savedSearch,
+      location,
     });
-  }
+  }, [getQueryFromSavedSearchOrLocation, savedSearch, location]);
 
-  getSort(): string {
-    return this.getSortFromSavedSearchOrLocation({
-      savedSearch: this.props.savedSearch,
-      location: this.props.location,
+  const sort = useMemo((): string => {
+    return getSortFromSavedSearchOrLocation({
+      savedSearch,
+      location,
     });
-  }
+  }, [getSortFromSavedSearchOrLocation, savedSearch, location]);
 
-  getGroupStatsPeriod(): string {
+  const getGroupStatsPeriod = useCallback((): string => {
     let currentPeriod: string;
-    if (typeof this.props.location.query?.groupStatsPeriod === 'string') {
-      currentPeriod = this.props.location.query.groupStatsPeriod;
+    if (typeof location.query?.groupStatsPeriod === 'string') {
+      currentPeriod = location.query.groupStatsPeriod;
     } else {
       currentPeriod = DEFAULT_GRAPH_STATS_PERIOD;
     }
@@ -390,15 +263,13 @@ class IssueListOverview extends Component<Props, State> {
     return DYNAMIC_COUNTS_STATS_PERIODS.has(currentPeriod)
       ? currentPeriod
       : DEFAULT_GRAPH_STATS_PERIOD;
-  }
+  }, [location]);
 
-  getEndpointParams = (): EndpointParams => {
-    const {selection} = this.props;
-
+  const getEndpointParams = useCallback((): EndpointParams => {
     const params: EndpointParams = {
       project: selection.projects,
       environment: selection.environments,
-      query: this.getQuery(),
+      query,
       ...selection.datetime,
     };
 
@@ -413,78 +284,228 @@ class IssueListOverview extends Component<Props, State> {
       params.start = getUtcDateString(params.start);
     }
 
-    const sort = this.getSort();
     if (sort !== DEFAULT_ISSUE_STREAM_SORT) {
       params.sort = sort;
     }
 
-    const groupStatsPeriod = this.getGroupStatsPeriod();
+    const groupStatsPeriod = getGroupStatsPeriod();
     if (groupStatsPeriod !== DEFAULT_GRAPH_STATS_PERIOD) {
       params.groupStatsPeriod = groupStatsPeriod;
     }
 
-    if (this.props.location.query.useGroupSnubaDataset) {
+    if (location.query.useGroupSnubaDataset) {
       params.useGroupSnubaDataset = true;
     }
 
     // only include defined values.
     return pickBy(params, v => defined(v)) as EndpointParams;
-  };
+  }, [selection, location, query, sort, getGroupStatsPeriod]);
 
-  getCacheEndpointParams = (): EndpointParams => {
-    const cursor = this.props.location.query.cursor;
-    return {
-      ...this.getEndpointParams(),
-      cursor,
+  const requestParams = useMemo(() => {
+    // Used for Issue Stream Performance project, enabled means we are doing saved search look up in the backend
+    const savedSearchLookupEnabled = 0;
+    const savedSearchLookupDisabled = 1;
+
+    const params: any = {
+      ...getEndpointParams(),
+      limit: MAX_ITEMS,
+      shortIdLookup: 1,
+      savedSearch: savedSearchLoading
+        ? savedSearchLookupEnabled
+        : savedSearchLookupDisabled,
     };
-  };
 
-  getSelectedProjectIds = (): string[] => {
-    return this.props.selection.projects.map(projectId => String(projectId));
-  };
+    if (selectedSearchId) {
+      params.searchId = selectedSearchId;
+    }
 
-  fetchMemberList() {
-    const projectIds = this.getSelectedProjectIds();
+    if (savedSearchLoading && !defined(location.query.query)) {
+      delete params.query;
+    }
 
-    fetchOrgMembers(this.props.api, this.props.organization.slug, projectIds).then(
-      members => {
-        this.setState({memberList: indexMembersByProject(members)});
-      }
-    );
-  }
+    const currentQuery = location.query || {};
+    if ('cursor' in currentQuery) {
+      params.cursor = currentQuery.cursor;
+    }
 
-  fetchStats = (groups: string[]) => {
-    // If we have no groups to fetch, just skip stats
-    if (!groups.length) {
+    // If no stats period values are set, use default
+    if (!params.statsPeriod && !params.start) {
+      params.statsPeriod = DEFAULT_STATS_PERIOD;
+    }
+
+    params.expand = ['owners', 'inbox'];
+    params.collapse = ['stats', 'unhandled'];
+
+    return params;
+  }, [getEndpointParams, location.query, savedSearchLoading, selectedSearchId]);
+
+  const loadFromCache = useCallback((): boolean => {
+    const cache = IssueListCacheStore.getFromCache(requestParams);
+
+    if (!cache) {
+      return false;
+    }
+
+    setIssuesLoading(false);
+    setQueryCount(cache.queryCount);
+    setQueryMaxCount(cache.queryMaxCount);
+    setPageLinks(cache.pageLinks);
+
+    GroupStore.add(cache.groups);
+
+    return true;
+  }, [requestParams]);
+
+  const resumePolling = useCallback(() => {
+    if (!pageLinks) {
       return;
     }
-    const requestParams: StatEndpointParams = {
-      ...this.getEndpointParams(),
-      groups,
-    };
-    // If no stats period values are set, use default
-    if (!requestParams.statsPeriod && !requestParams.start) {
-      requestParams.statsPeriod = DEFAULT_STATS_PERIOD;
+
+    // Only resume polling if we're on the first page of results
+    const links = parseLinkHeader(pageLinks);
+    if (links && !links.previous!.results && realtimeActive) {
+      pollerRef.current?.setEndpoint(links?.previous!.href);
+      pollerRef.current?.enable();
     }
+  }, [pageLinks, realtimeActive]);
 
-    this._lastStatsRequest = this.props.api.request(this.groupStatsEndpoint, {
-      method: 'GET',
-      data: qs.stringify(requestParams),
-      success: data => {
-        if (!data) {
-          return;
+  const trackTabViewed = useCallback(
+    (newGroupIds: string[], data: Group[], numHits: number | null) => {
+      const endpointParams = getEndpointParams();
+      const tabQueriesWithCounts = getTabsWithCounts();
+      const currentTabQuery = tabQueriesWithCounts.includes(endpointParams.query as Query)
+        ? endpointParams.query
+        : null;
+      const tab = getTabs().find(([tabQuery]) => currentTabQuery === tabQuery)?.[1];
+
+      const numPerfIssues = newGroupIds.filter(
+        groupId => GroupStore.get(groupId)?.issueCategory === IssueCategory.PERFORMANCE
+      ).length;
+      // First and last seen are only available after the group has fetched stats
+      // Number of issues shown whose first seen is more than 30 days ago
+      const numOldIssues = data.filter((group: BaseGroup) =>
+        moment(new Date(group.firstSeen)).isBefore(moment().subtract(30, 'd'))
+      ).length;
+      // number of issues shown whose first seen is less than 7 days
+      const numNewIssues = data.filter((group: BaseGroup) =>
+        moment(new Date(group.firstSeen)).isAfter(moment().subtract(7, 'd'))
+      ).length;
+
+      trackAnalytics('issues_tab.viewed', {
+        organization,
+        tab: tab?.analyticsName,
+        page: parsePageQueryParam(location, 0),
+        query,
+        num_perf_issues: numPerfIssues,
+        num_old_issues: numOldIssues,
+        num_new_issues: numNewIssues,
+        num_issues: data.length,
+        total_issues_count: numHits,
+        issue_views_enabled: organization.features.includes('issue-stream-custom-views'),
+        sort,
+      });
+    },
+    [organization, location, getEndpointParams, query, sort]
+  );
+
+  const fetchCounts = useCallback(
+    (currentQueryCount: number, fetchAllCounts: boolean) => {
+      let newQueryCounts: QueryCounts = {...queryCounts};
+
+      const endpointParams = getEndpointParams();
+      const tabQueriesWithCounts = getTabsWithCounts();
+      const currentTabQuery = tabQueriesWithCounts.includes(endpointParams.query as Query)
+        ? endpointParams.query
+        : null;
+
+      // Update the count based on the exact number of issues, these shown as is
+      if (currentTabQuery) {
+        // @ts-expect-error TS(7053): Element implicitly has an 'any' type because expre... Remove this comment to see the full error message
+        newQueryCounts[currentTabQuery] = {
+          count: currentQueryCount,
+          hasMore: false,
+        };
+      }
+
+      setQueryCounts(newQueryCounts);
+
+      // If all tabs' counts are fetched, skip and only set
+      if (
+        fetchAllCounts ||
+        // @ts-expect-error TS(7053): Element implicitly has an 'any' type because expre... Remove this comment to see the full error message
+        !tabQueriesWithCounts.every(tabQuery => queryCounts[tabQuery] !== undefined)
+      ) {
+        const countsRequestParams: CountsEndpointParams = {
+          ...omit(endpointParams, 'query'),
+          // fetch the counts for the tabs whose counts haven't been fetched yet
+          query: tabQueriesWithCounts.filter(_query => _query !== currentTabQuery),
+        };
+
+        // If no stats period values are set, use default
+        if (!countsRequestParams.statsPeriod && !countsRequestParams.start) {
+          countsRequestParams.statsPeriod = DEFAULT_STATS_PERIOD;
         }
-        GroupStore.onPopulateStats(groups, data);
-        this.trackTabViewed(groups, data, this.state.queryCount);
-      },
-      error: err => {
-        this.setState({
-          error: parseApiError(err),
-        });
-      },
-      complete: () => {
-        this._lastStatsRequest = null;
 
+        api.request(`/organizations/${organization.slug}/issues-count/`, {
+          method: 'GET',
+          data: qs.stringify(countsRequestParams),
+
+          success: data => {
+            if (!data) {
+              return;
+            }
+            // Counts coming from the counts endpoint is limited to 100, for >= 100 we display 99+
+            newQueryCounts = {
+              ...queryCounts,
+              ...mapValues(data, (count: number) => ({
+                count,
+                hasMore: count > TAB_MAX_COUNT,
+              })),
+            };
+          },
+          error: () => {
+            setQueryCounts({});
+          },
+          complete: () => {
+            setQueryCounts(newQueryCounts);
+          },
+        });
+      }
+    },
+    [api, getEndpointParams, organization.slug, queryCounts]
+  );
+
+  const fetchStats = useCallback(
+    async (newGroupIds: string[]) => {
+      // If we have no groups to fetch, just skip stats
+      if (!newGroupIds.length) {
+        return;
+      }
+      const statsRequestParams: StatEndpointParams = {
+        ...getEndpointParams(),
+        groups: newGroupIds,
+      };
+      // If no stats period values are set, use default
+      if (!statsRequestParams.statsPeriod && !statsRequestParams.start) {
+        statsRequestParams.statsPeriod = DEFAULT_STATS_PERIOD;
+      }
+
+      try {
+        const data = await api.requestPromise(
+          `/organizations/${organization.slug}/issues-stats/`,
+          {
+            method: 'GET',
+            data: qs.stringify(statsRequestParams),
+          }
+        );
+
+        if (data) {
+          GroupStore.onPopulateStats(newGroupIds, data);
+          trackTabViewed(newGroupIds, data, queryCount);
+        }
+      } catch (e) {
+        setError(parseApiError(e));
+      } finally {
         // End navigation transaction to prevent additional page requests from impacting page metrics.
         // Other transactions include stacktrace preview request
         const currentSpan = Sentry.getActiveSpan();
@@ -492,368 +513,348 @@ class IssueListOverview extends Component<Props, State> {
         if (rootSpan && Sentry.spanToJSON(rootSpan).op === 'navigation') {
           rootSpan.end();
         }
-      },
-    });
-  };
+      }
+    },
+    [getEndpointParams, api, organization.slug, trackTabViewed, queryCount]
+  );
 
-  fetchCounts = (currentQueryCount: number, fetchAllCounts: boolean) => {
-    const {queryCounts: _queryCounts} = this.state;
-    let queryCounts: QueryCounts = {..._queryCounts};
+  const fetchData = useCallback(
+    (fetchAllCounts = false) => {
+      if (realtimeActive || (!actionTakenRef.current && !undoRef.current)) {
+        GroupStore.loadInitialData([]);
 
-    const endpointParams = this.getEndpointParams();
-    const tabQueriesWithCounts = getTabsWithCounts();
-    const currentTabQuery = tabQueriesWithCounts.includes(endpointParams.query as Query)
-      ? endpointParams.query
-      : null;
-
-    // Update the count based on the exact number of issues, these shown as is
-    if (currentTabQuery) {
-      // @ts-expect-error TS(7053): Element implicitly has an 'any' type because expre... Remove this comment to see the full error message
-      queryCounts[currentTabQuery] = {
-        count: currentQueryCount,
-        hasMore: false,
-      };
-    }
-
-    this.setState({queryCounts});
-
-    // If all tabs' counts are fetched, skip and only set
-    if (
-      fetchAllCounts ||
-      // @ts-expect-error TS(7053): Element implicitly has an 'any' type because expre... Remove this comment to see the full error message
-      !tabQueriesWithCounts.every(tabQuery => queryCounts[tabQuery] !== undefined)
-    ) {
-      const requestParams: CountsEndpointParams = {
-        ...omit(endpointParams, 'query'),
-        // fetch the counts for the tabs whose counts haven't been fetched yet
-        query: tabQueriesWithCounts.filter(_query => _query !== currentTabQuery),
-      };
-
-      // If no stats period values are set, use default
-      if (!requestParams.statsPeriod && !requestParams.start) {
-        requestParams.statsPeriod = DEFAULT_STATS_PERIOD;
+        setIssuesLoading(true);
+        setQueryCount(0);
+        setError(null);
       }
 
-      this._lastFetchCountsRequest = this.props.api.request(this.groupCountsEndpoint, {
+      const span = getCurrentSentryReactRootSpan();
+      span?.setAttribute('query.sort', sort);
+
+      setError(null);
+
+      api.clear();
+      pollerRef.current?.disable();
+
+      api.request(`/organizations/${organization.slug}/issues/`, {
         method: 'GET',
         data: qs.stringify(requestParams),
-
-        success: data => {
-          if (!data) {
+        success: async (data, _, resp) => {
+          if (!resp) {
             return;
           }
-          // Counts coming from the counts endpoint is limited to 100, for >= 100 we display 99+
-          queryCounts = {
-            ...queryCounts,
-            ...mapValues(data, (count: number) => ({
-              count,
-              hasMore: count > TAB_MAX_COUNT,
-            })),
-          };
-        },
-        error: () => {
-          this.setState({queryCounts: {}});
-        },
-        complete: () => {
-          this._lastFetchCountsRequest = null;
 
-          this.setState({queryCounts});
-        },
-      });
-    }
-  };
+          // If this is a direct hit, we redirect to the intended result directly.
+          if (resp.getResponseHeader('X-Sentry-Direct-Hit') === '1') {
+            let redirect: string;
+            if (data[0]?.matchingEventId) {
+              const {id, matchingEventId} = data[0];
+              redirect = `/organizations/${organization.slug}/issues/${id}/events/${matchingEventId}/`;
+            } else {
+              const {id} = data[0];
+              redirect = `/organizations/${organization.slug}/issues/${id}/`;
+            }
 
-  fetchData = (fetchAllCounts = false) => {
-    const {organization} = this.props;
-    const query = this.getQuery();
-
-    if (this.state.realtimeActive || (!this.actionTaken && !this.undo)) {
-      GroupStore.loadInitialData([]);
-
-      this.setState({
-        issuesLoading: true,
-        queryCount: 0,
-        error: null,
-      });
-    }
-
-    const span = getCurrentSentryReactRootSpan();
-    span?.setAttribute('query.sort', this.getSort());
-
-    this.setState({
-      error: null,
-    });
-
-    // Used for Issue Stream Performance project, enabled means we are doing saved search look up in the backend
-    const savedSearchLookupEnabled = 0;
-    const savedSearchLookupDisabled = 1;
-
-    const requestParams: any = {
-      ...this.getEndpointParams(),
-      limit: MAX_ITEMS,
-      shortIdLookup: 1,
-      savedSearch: this.props.organization.features.includes('issue-stream-performance')
-        ? this.props.savedSearchLoading
-          ? savedSearchLookupEnabled
-          : savedSearchLookupDisabled
-        : savedSearchLookupDisabled,
-    };
-
-    if (
-      this.props.organization.features.includes('issue-stream-performance') &&
-      this.props.selectedSearchId
-    ) {
-      requestParams.searchId = this.props.selectedSearchId;
-    }
-
-    if (
-      this.props.organization.features.includes('issue-stream-performance') &&
-      this.props.savedSearchLoading &&
-      !defined(this.props.location.query.query)
-    ) {
-      delete requestParams.query;
-    }
-
-    const currentQuery = this.props.location.query || {};
-    if ('cursor' in currentQuery) {
-      requestParams.cursor = currentQuery.cursor;
-    }
-
-    // If no stats period values are set, use default
-    if (!requestParams.statsPeriod && !requestParams.start) {
-      requestParams.statsPeriod = DEFAULT_STATS_PERIOD;
-    }
-
-    requestParams.expand = ['owners', 'inbox'];
-    requestParams.collapse = ['stats', 'unhandled'];
-
-    if (this._lastRequest) {
-      this._lastRequest.cancel();
-    }
-    if (this._lastStatsRequest) {
-      this._lastStatsRequest.cancel();
-    }
-    if (this._lastFetchCountsRequest) {
-      this._lastFetchCountsRequest.cancel();
-    }
-
-    this._poller.disable();
-
-    this._lastRequest = this.props.api.request(this.groupListEndpoint, {
-      method: 'GET',
-      data: qs.stringify(requestParams),
-      success: (data, _, resp) => {
-        if (!resp) {
-          return;
-        }
-
-        // If this is a direct hit, we redirect to the intended result directly.
-        if (resp.getResponseHeader('X-Sentry-Direct-Hit') === '1') {
-          let redirect: string;
-          if (data[0]?.matchingEventId) {
-            const {id, matchingEventId} = data[0];
-            redirect = `/organizations/${organization.slug}/issues/${id}/events/${matchingEventId}/`;
-          } else {
-            const {id} = data[0];
-            redirect = `/organizations/${organization.slug}/issues/${id}/`;
+            navigate(
+              normalizeUrl({
+                pathname: redirect,
+                query: {
+                  referrer: 'issue-list',
+                  ...extractSelectionParameters(location.query),
+                },
+              }),
+              {replace: true}
+            );
+            return;
           }
 
-          browserHistory.replace(
-            normalizeUrl({
-              pathname: redirect,
-              query: {
-                referrer: 'issue-list',
-                ...extractSelectionParameters(this.props.location.query),
-              },
-            })
-          );
-          return;
-        }
+          if (undoRef.current) {
+            GroupStore.loadInitialData(data);
+          }
+          GroupStore.add(data);
 
-        if (this.undo) {
-          GroupStore.loadInitialData(data);
-        }
-        GroupStore.add(data);
+          await fetchStats(data.map((group: BaseGroup) => group.id));
 
-        this.fetchStats(data.map((group: BaseGroup) => group.id));
+          const hits = resp.getResponseHeader('X-Hits');
+          const newQueryCount =
+            typeof hits !== 'undefined' && hits ? parseInt(hits, 10) || 0 : 0;
+          const maxHits = resp.getResponseHeader('X-Max-Hits');
+          const newQueryMaxCount =
+            typeof maxHits !== 'undefined' && maxHits ? parseInt(maxHits, 10) || 0 : 0;
+          const newPageLinks = resp.getResponseHeader('Link');
 
-        const hits = resp.getResponseHeader('X-Hits');
-        const queryCount =
-          typeof hits !== 'undefined' && hits ? parseInt(hits, 10) || 0 : 0;
-        const maxHits = resp.getResponseHeader('X-Max-Hits');
-        const queryMaxCount =
-          typeof maxHits !== 'undefined' && maxHits ? parseInt(maxHits, 10) || 0 : 0;
-        const pageLinks = resp.getResponseHeader('Link');
+          fetchCounts(newQueryCount, fetchAllCounts);
 
-        this.fetchCounts(queryCount, fetchAllCounts);
+          setError(null);
+          setIssuesLoading(false);
+          setQueryCount(newQueryCount);
+          setQueryMaxCount(newQueryMaxCount);
+          setPageLinks(newPageLinks !== null ? newPageLinks : '');
 
-        this.setState({
-          error: null,
-          issuesLoading: false,
-          queryCount,
-          queryMaxCount,
-          pageLinks: pageLinks !== null ? pageLinks : '',
-        });
+          IssueListCacheStore.save(requestParams, {
+            groups: GroupStore.getState() as Group[],
+            queryCount: newQueryCount,
+            queryMaxCount: newQueryMaxCount,
+            pageLinks: newPageLinks ?? '',
+          });
 
-        if (data.length === 0) {
-          trackAnalytics('issue_search.empty', {
-            organization: this.props.organization,
+          if (data.length === 0) {
+            trackAnalytics('issue_search.empty', {
+              organization,
+              search_type: 'issues',
+              search_source: 'main_search',
+              query,
+            });
+          }
+        },
+        error: err => {
+          trackAnalytics('issue_search.failed', {
+            organization,
             search_type: 'issues',
             search_source: 'main_search',
-            query,
+            error: parseApiError(err),
           });
-        }
-      },
-      error: err => {
-        trackAnalytics('issue_search.failed', {
-          organization: this.props.organization,
-          search_type: 'issues',
-          search_source: 'main_search',
-          error: parseApiError(err),
-        });
 
-        this.setState({
-          error: parseApiError(err),
-          issuesLoading: false,
-        });
-      },
-      complete: () => {
-        this._lastRequest = null;
+          setError(parseApiError(err));
+          setIssuesLoading(false);
+        },
+        complete: () => {
+          resumePolling();
 
-        this.resumePolling();
+          if (!realtimeActive) {
+            actionTakenRef.current = false;
+            undoRef.current = false;
+          }
+        },
+      });
+    },
+    [
+      realtimeActive,
+      sort,
+      api,
+      organization,
+      requestParams,
+      fetchStats,
+      fetchCounts,
+      navigate,
+      location.query,
+      query,
+      resumePolling,
+    ]
+  );
 
-        if (!this.state.realtimeActive) {
-          this.actionTaken = false;
-          this.undo = false;
-        }
-      },
-    });
-  };
+  useRouteAnalyticsParams({
+    issue_views_enabled: organization.features.includes('issue-stream-custom-views'),
+  });
+  useDisableRouteAnalytics();
 
-  resumePolling = () => {
-    if (!this.state.pageLinks) {
+  // Update polling status
+  useEffect(() => {
+    if (realtimeActive) {
+      resumePolling();
+    } else {
+      pollerRef.current?.disable();
+    }
+  }, [realtimeActive, resumePolling]);
+
+  // Fetch data on mount if necessary
+  useEffect(() => {
+    const loadedFromCache = loadFromCache();
+    if (!loadedFromCache) {
+      // It's possible the projects query parameter is not yet ready and this
+      // request will be repeated in componentDidUpdate
+      fetchData();
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  const previousSelection = usePrevious(selection);
+  const previousSavedSearchLoading = usePrevious(savedSearchLoading);
+  const previousIssuesLoading = usePrevious(issuesLoading);
+
+  const previousRequestParams = usePrevious(requestParams);
+
+  // Keep data up to date
+  useEffect(() => {
+    const selectionChanged = !isEqual(previousSelection, selection);
+
+    // Wait for saved searches to load before we attempt to fetch stream data
+    // Selection changing could indicate that the projects query parameter has populated
+    // and we should refetch data.
+    if (savedSearchLoading && !selectionChanged) {
       return;
     }
 
-    // Only resume polling if we're on the first page of results
-    const links = parseLinkHeader(this.state.pageLinks);
-    if (links && !links.previous!.results && this.state.realtimeActive) {
-      this._poller.setEndpoint(links?.previous?.href);
-      this._poller.enable();
+    if (previousSavedSearchLoading && !savedSearchLoading) {
+      return;
     }
-  };
 
-  get groupListEndpoint(): string {
-    const {organization} = this.props;
-    return `/organizations/${organization.slug}/issues/`;
-  }
+    // If any important url parameter changed or saved search changed
+    // reload data.
+    if (!isEqual(previousRequestParams, requestParams)) {
+      fetchData(selectionChanged);
+    }
+  }, [
+    fetchData,
+    savedSearchLoading,
+    selection,
+    previousSelection,
+    organization.features,
+    issuesLoading,
+    loadFromCache,
+    previousSavedSearchLoading,
+    previousIssuesLoading,
+    previousRequestParams,
+    requestParams,
+  ]);
 
-  get groupCountsEndpoint(): string {
-    const {organization} = this.props;
-    return `/organizations/${organization.slug}/issues-count/`;
-  }
+  // Fetch members on mount
+  useEffect(() => {
+    const projectIds = selection.projects.map(projectId => String(projectId));
 
-  get groupStatsEndpoint(): string {
-    const {organization} = this.props;
-    return `/organizations/${organization.slug}/issues-stats/`;
-  }
-
-  onRealtimeChange = (realtime: boolean) => {
-    Cookies.set('realtimeActive', realtime.toString());
-    this.setState({realtimeActive: realtime});
-    trackAnalytics('issues_stream.realtime_clicked', {
-      organization: this.props.organization,
-      enabled: realtime,
+    fetchOrgMembers(api, organization.slug, projectIds).then(members => {
+      setMemberList(indexMembersByProject(members));
     });
-  };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
-  onSelectStatsPeriod = (period: string) => {
-    const {location} = this.props;
-    if (period !== this.getGroupStatsPeriod()) {
-      const cursor = location.query.cursor;
-      const queryPageInt = parseInt(location.query.page, 10);
-      const page = isNaN(queryPageInt) || !location.query.cursor ? 0 : queryPageInt;
-      this.transitionTo({cursor, page, groupStatsPeriod: period});
+  // If the project selection has changed reload the member list and tag keys
+  // allowing autocomplete and tag sidebar to be more accurate.
+  useEffect(() => {
+    if (isEqual(previousSelection?.projects, selection.projects)) {
+      return;
     }
-  };
 
-  onRealtimePoll = (data: any, {queryCount}: {queryCount: number}) => {
-    // Note: We do not update state with cursors from polling,
-    // `CursorPoller` updates itself with new cursors
-    GroupStore.addToFront(data);
-    this.setState({queryCount});
-  };
+    const projectIds = selection.projects.map(projectId => String(projectId));
 
-  listener = GroupStore.listen(() => this.onGroupChange(), undefined);
-
-  onGroupChange() {
-    const groupIds = GroupStore.getAllItems()
-      .map(item => item.id)
-      .slice(0, MAX_ISSUES_COUNT);
-
-    if (!isEqual(groupIds, this.state.groupIds)) {
-      this.setState({groupIds});
-    }
-  }
-
-  trackTabViewed(groups: string[], data: Group[], numHits: number | null) {
-    const {organization, location} = this.props;
-    const page = location.query.page;
-    const endpointParams = this.getEndpointParams();
-    const tabQueriesWithCounts = getTabsWithCounts();
-    const currentTabQuery = tabQueriesWithCounts.includes(endpointParams.query as Query)
-      ? endpointParams.query
-      : null;
-    const tab = getTabs().find(([tabQuery]) => currentTabQuery === tabQuery)?.[1];
-
-    const numPerfIssues = groups.filter(
-      group => GroupStore.get(group)?.issueCategory === IssueCategory.PERFORMANCE
-    ).length;
-    // First and last seen are only available after the group has fetched stats
-    // Number of issues shown whose first seen is more than 30 days ago
-    const numOldIssues = data.filter((group: BaseGroup) =>
-      moment(new Date(group.firstSeen)).isBefore(moment().subtract(30, 'd'))
-    ).length;
-    // number of issues shown whose first seen is less than 7 days
-    const numNewIssues = data.filter((group: BaseGroup) =>
-      moment(new Date(group.firstSeen)).isAfter(moment().subtract(7, 'd'))
-    ).length;
-
-    trackAnalytics('issues_tab.viewed', {
-      organization,
-      tab: tab?.analyticsName,
-      page: page ? parseInt(page, 10) : 0,
-      query: this.getQuery(),
-      num_perf_issues: numPerfIssues,
-      num_old_issues: numOldIssues,
-      num_new_issues: numNewIssues,
-      num_issues: data.length,
-      total_issues_count: numHits,
-      issue_views_enabled: organization.features.includes('issue-stream-custom-views'),
-      sort: this.getSort(),
+    fetchOrgMembers(api, organization.slug, projectIds).then(members => {
+      setMemberList(indexMembersByProject(members));
     });
-  }
+  }, [api, organization.slug, selection.projects, previousSelection?.projects]);
 
-  onSearch = (query: string) => {
-    if (query === this.state.query) {
+  // Cleanup
+  useEffect(() => {
+    return () => {
+      pollerRef.current?.disable();
+      SelectedGroupStore.reset();
+      GroupStore.reset();
+    };
+  }, []);
+
+  const allResultsVisible = useCallback(() => {
+    if (!pageLinks) {
+      return false;
+    }
+
+    const links = parseLinkHeader(pageLinks);
+    return links && !links.previous!.results && !links.next!.results;
+  }, [pageLinks]);
+
+  const getPageCounts = useCallback(() => {
+    const links = parseLinkHeader(pageLinks);
+    const queryPageInt = parsePageQueryParam(location, 0);
+    // Cursor must be present for the page number to be used
+    const page = !location.query.cursor ? 0 : queryPageInt;
+
+    let numPreviousIssues = Math.min(page * MAX_ITEMS, queryCount);
+
+    // Because the query param `page` is not tied to the request, we need to
+    // validate that it's correct at the first and last page
+    if (!links?.next?.results || allResultsVisible()) {
+      // On last available page
+      numPreviousIssues = Math.max(queryCount - groupIds.length, 0);
+    } else if (!links?.previous?.results) {
+      // On first available page
+      numPreviousIssues = 0;
+    }
+
+    return {
+      numPreviousIssues,
+      numIssuesOnPage: groupIds.length,
+    };
+  }, [pageLinks, location, queryCount, allResultsVisible, groupIds.length]);
+
+  const onRealtimeChange = useCallback(
+    (realtime: boolean) => {
+      Cookies.set('realtimeActive', realtime.toString());
+      setRealtimeActive(realtime);
+      trackAnalytics('issues_stream.realtime_clicked', {
+        organization,
+        enabled: realtime,
+      });
+    },
+    [organization]
+  );
+
+  const transitionTo = (
+    newParams: Partial<EndpointParams> = {},
+    newSavedSearch: (SavedSearch & {projectId?: number}) | null = savedSearch
+  ) => {
+    const queryData = {
+      ...omit(location.query, ['page', 'cursor']),
+      referrer: 'issue-list',
+      ...getEndpointParams(),
+      ...newParams,
+    };
+    let path: string;
+
+    if (newSavedSearch?.id) {
+      path = `/organizations/${organization.slug}/issues/searches/${newSavedSearch.id}/`;
+
+      // Remove the query as saved searches bring their own query string.
+      delete queryData.query;
+
+      // If we aren't going to another page in the same search
+      // drop the query and replace the current project, with the saved search search project
+      // if available.
+      if (!queryData.cursor && newSavedSearch.projectId) {
+        queryData.project = [newSavedSearch.projectId];
+      }
+      if (!queryData.cursor && !newParams.sort && newSavedSearch.sort) {
+        queryData.sort = newSavedSearch.sort;
+      }
+    } else {
+      path = `/organizations/${organization.slug}/issues/`;
+    }
+
+    if (
+      queryData.sort === IssueSortOptions.INBOX &&
+      !FOR_REVIEW_QUERIES.includes(queryData.query || '')
+    ) {
+      delete queryData.sort;
+    }
+
+    if (path !== location.pathname || !isEqual(query, location.query)) {
+      navigate({
+        pathname: normalizeUrl(path),
+        query: queryData,
+      });
+      setIssuesLoading(true);
+    }
+  };
+
+  const onSearch = (newQuery: string) => {
+    if (newQuery === query) {
       // if query is the same, just re-fetch data
-      this.fetchData();
+      fetchData();
     } else {
       // Clear the saved search as the user wants something else.
-      this.transitionTo({query}, null);
+      transitionTo({query: newQuery}, null);
     }
   };
 
-  onSortChange = (sort: string) => {
+  const onSortChange = (newSort: string) => {
     trackAnalytics('issues_stream.sort_changed', {
-      organization: this.props.organization,
-      sort,
+      organization,
+      sort: newSort,
     });
-    this.transitionTo({sort});
+    transitionTo({sort: newSort});
   };
 
-  onCursorChange: CursorHandler = (nextCursor, _path, _query, delta) => {
-    const queryPageInt = parseInt(this.props.location.query.page, 10);
+  const onCursorChange: CursorHandler = (nextCursor, _path, _query, delta) => {
+    const queryPageInt = Array.isArray(location.query.page)
+      ? NaN
+      : parseInt(location.query.page?.toString() ?? '', 10);
     let nextPage: number | undefined = isNaN(queryPageInt) ? delta : queryPageInt + delta;
 
     let cursor: undefined | string = nextCursor;
@@ -866,135 +867,38 @@ class IssueListOverview extends Component<Props, State> {
       nextPage = undefined;
     }
 
-    this.transitionTo({cursor, page: nextPage});
+    transitionTo({cursor, page: nextPage});
   };
 
-  paginationAnalyticsEvent = (direction: string) => {
-    trackAnalytics('issues_stream.paginate', {
-      organization: this.props.organization,
-      direction,
-    });
-  };
-
-  /**
-   * Returns true if all results in the current query are visible/on this page
-   */
-  allResultsVisible(): boolean {
-    if (!this.state.pageLinks) {
-      return false;
-    }
-
-    const links = parseLinkHeader(this.state.pageLinks);
-    return links && !links.previous!.results && !links.next!.results;
-  }
-
-  transitionTo = (
-    newParams: Partial<EndpointParams> = {},
-    savedSearch: (SavedSearch & {projectId?: number}) | null = this.props.savedSearch
-  ) => {
-    const query = {
-      ...omit(this.props.location.query, ['page', 'cursor']),
-      referrer: 'issue-list',
-      ...this.getEndpointParams(),
-      ...newParams,
-    };
-    const {organization} = this.props;
-    let path: string;
-
-    if (savedSearch?.id) {
-      path = `/organizations/${organization.slug}/issues/searches/${savedSearch.id}/`;
-
-      // Remove the query as saved searches bring their own query string.
-      delete query.query;
-
-      // If we aren't going to another page in the same search
-      // drop the query and replace the current project, with the saved search search project
-      // if available.
-      if (!query.cursor && savedSearch.projectId) {
-        query.project = [savedSearch.projectId];
-      }
-      if (!query.cursor && !newParams.sort && savedSearch.sort) {
-        query.sort = savedSearch.sort;
-      }
-    } else {
-      path = `/organizations/${organization.slug}/issues/`;
-    }
-
-    if (
-      query.sort === IssueSortOptions.INBOX &&
-      !FOR_REVIEW_QUERIES.includes(query.query || '')
-    ) {
-      delete query.sort;
-    }
-
-    if (
-      path !== this.props.location.pathname ||
-      !isEqual(query, this.props.location.query)
-    ) {
-      browserHistory.push({
-        pathname: normalizeUrl(path),
-        query,
-      });
-      this.setState({issuesLoading: true});
+  const onSelectStatsPeriod = (period: string) => {
+    if (period !== getGroupStatsPeriod()) {
+      const cursor = Array.isArray(location.query.cursor)
+        ? location.query.cursor[0]
+        : location.query.cursor ?? undefined;
+      const queryPageInt = parsePageQueryParam(location, 0);
+      const page = !location.query.cursor ? 0 : queryPageInt;
+      transitionTo({cursor, page, groupStatsPeriod: period});
     }
   };
 
-  displayReprocessingTab() {
-    return !!this.state.queryCounts?.[Query.REPROCESSING]?.count;
-  }
-
-  displayReprocessingLayout(showReprocessingTab: boolean, query: string) {
-    return showReprocessingTab && query === Query.REPROCESSING;
-  }
-
-  renderLoading(): React.ReactNode {
-    return (
-      <Layout.Page withPadding>
-        <LoadingIndicator />
-      </Layout.Page>
-    );
-  }
-
-  onSavedSearchSelect = (savedSearch: SavedSearch) => {
-    trackAnalytics('organization_saved_search.selected', {
-      organization: this.props.organization,
-      search_type: 'issues',
-      id: savedSearch.id ? parseInt(savedSearch.id, 10) : -1,
-      is_global: savedSearch.isGlobal,
-      query: savedSearch.query,
-      visibility: savedSearch.visibility,
-    });
-    this.setState({issuesLoading: true}, () => this.transitionTo(undefined, savedSearch));
-  };
-
-  onDelete = () => {
-    this.actionTaken = true;
-    this.fetchData(true);
-  };
-
-  undoAction = ({data, groups}: {data: IssueUpdateData; groups: BaseGroup[]}) => {
-    const {organization, selection} = this.props;
-    const query = this.getQuery();
-
+  const undoAction = ({
+    data,
+    groupItems,
+  }: {
+    data: IssueUpdateData;
+    groupItems: BaseGroup[];
+  }) => {
     const projectIds = selection?.projects?.map(p => p.toString());
     const endpoint = `/organizations/${organization.slug}/issues/`;
 
-    if (this._lastRequest) {
-      this._lastRequest.cancel();
-    }
-    if (this._lastStatsRequest) {
-      this._lastStatsRequest.cancel();
-    }
-    if (this._lastFetchCountsRequest) {
-      this._lastFetchCountsRequest.cancel();
-    }
+    api.clear();
 
-    this.props.api.request(endpoint, {
+    api.request(endpoint, {
       method: 'PUT',
       data,
       query: {
         project: projectIds,
-        id: groups.map(group => group.id),
+        id: groupItems.map(group => group.id),
       },
       success: response => {
         if (!response) {
@@ -1004,86 +908,21 @@ class IssueListOverview extends Component<Props, State> {
         // on this page for a second and then be removed (will show up on All Unresolved). This is to
         // stop this from happening and avoid confusion.
         if (!query.includes('is:ignored') && !isForReviewQuery(query)) {
-          GroupStore.add(groups);
+          GroupStore.add(groupItems);
         }
-        this.undo = true;
+        actionTakenRef.current = true;
       },
       error: err => {
-        this.setState({
-          error: parseApiError(err),
-          issuesLoading: false,
-        });
+        setError(parseApiError(err));
+        setIssuesLoading(false);
       },
       complete: () => {
-        this.fetchData();
+        fetchData(true);
       },
     });
   };
 
-  onActionTaken = (itemIds: string[], data: IssueUpdateData) => {
-    if (this.state.realtimeActive) {
-      return;
-    }
-
-    const query = this.getQuery();
-    const groups = itemIds.map(id => GroupStore.get(id)).filter(defined);
-
-    if ('status' in data) {
-      if (data.status === 'resolved') {
-        this.onIssueAction({
-          itemIds,
-          actionType: 'Resolved',
-          shouldRemove:
-            query.includes('is:unresolved') ||
-            query.includes('is:ignored') ||
-            isForReviewQuery(query),
-          undo: () =>
-            this.undoAction({
-              data: {status: GroupStatus.UNRESOLVED, statusDetails: {}},
-              groups,
-            }),
-        });
-        return;
-      }
-
-      if (data.status === 'ignored') {
-        this.onIssueAction({
-          itemIds,
-          actionType: 'Archived',
-          shouldRemove: query.includes('is:unresolved') || isForReviewQuery(query),
-          undo: () =>
-            this.undoAction({
-              data: {status: GroupStatus.UNRESOLVED, statusDetails: {}},
-              groups,
-            }),
-        });
-        return;
-      }
-    }
-
-    if ('inbox' in data && data.inbox === false) {
-      this.onIssueAction({
-        itemIds,
-        actionType: 'Reviewed',
-        shouldRemove: isForReviewQuery(query),
-      });
-      return;
-    }
-
-    if ('priority' in data && typeof data.priority === 'string') {
-      const priorityValues = parseIssuePrioritySearch(query);
-      const priority = data.priority.toLowerCase() as PriorityLevel;
-
-      this.onIssueAction({
-        itemIds,
-        actionType: 'Reprioritized',
-        shouldRemove: !priorityValues.has(priority),
-      });
-      return;
-    }
-  };
-
-  onIssueAction = ({
+  const onIssueAction = ({
     itemIds,
     actionType,
     shouldRemove,
@@ -1111,185 +950,215 @@ class IssueListOverview extends Component<Props, State> {
       return;
     }
 
-    const links = parseLinkHeader(this.state.pageLinks);
+    const links = parseLinkHeader(pageLinks);
 
     GroupStore.remove(itemIds);
 
-    const queryCount = this.state.queryCount - itemIds.length;
-    this.actionTaken = true;
-    this.setState({queryCount});
+    const newQueryCount = queryCount - itemIds.length;
+    actionTakenRef.current = true;
+    setQueryCount(newQueryCount);
 
     if (GroupStore.getAllItemIds().length === 0) {
       // If we run out of issues on the last page, navigate back a page to
       // avoid showing an empty state - if not on the last page, just show a spinner
       const shouldGoBackAPage = links?.previous?.results && !links?.next?.results;
-      this.transitionTo({cursor: shouldGoBackAPage ? links.previous!.cursor : undefined});
-      this.fetchCounts(queryCount, true);
+      transitionTo({cursor: shouldGoBackAPage ? links.previous!.cursor : undefined});
+      fetchCounts(newQueryCount, true);
     } else {
-      this.fetchData(true);
+      fetchData(true);
     }
   };
 
-  getPageCounts = () => {
-    const {location} = this.props;
-    const {pageLinks, queryCount, groupIds} = this.state;
-    const links = parseLinkHeader(pageLinks);
-    const queryPageInt = parseInt(location.query.page, 10);
-    // Cursor must be present for the page number to be used
-    const page = isNaN(queryPageInt) || !location.query.cursor ? 0 : queryPageInt;
-
-    let numPreviousIssues = Math.min(page * MAX_ITEMS, queryCount);
-
-    // Because the query param `page` is not tied to the request, we need to
-    // validate that it's correct at the first and last page
-    if (!links?.next?.results || this.allResultsVisible()) {
-      // On last available page
-      numPreviousIssues = Math.max(queryCount - groupIds.length, 0);
-    } else if (!links?.previous?.results) {
-      // On first available page
-      numPreviousIssues = 0;
+  const onActionTaken = (itemIds: string[], data: IssueUpdateData) => {
+    if (realtimeActive) {
+      return;
     }
 
-    return {
-      numPreviousIssues,
-      numIssuesOnPage: groupIds.length,
-    };
+    const groupItems = itemIds.map(id => GroupStore.get(id)).filter(defined);
+
+    if ('status' in data) {
+      if (data.status === 'resolved') {
+        onIssueAction({
+          itemIds,
+          actionType: 'Resolved',
+          shouldRemove:
+            query.includes('is:unresolved') ||
+            query.includes('is:ignored') ||
+            isForReviewQuery(query),
+          undo: () =>
+            undoAction({
+              data: {status: GroupStatus.UNRESOLVED, statusDetails: {}},
+              groupItems,
+            }),
+        });
+        return;
+      }
+
+      if (data.status === 'ignored') {
+        onIssueAction({
+          itemIds,
+          actionType: 'Archived',
+          shouldRemove: query.includes('is:unresolved') || isForReviewQuery(query),
+          undo: () =>
+            undoAction({
+              data: {status: GroupStatus.UNRESOLVED, statusDetails: {}},
+              groupItems,
+            }),
+        });
+        return;
+      }
+    }
+
+    if ('inbox' in data && data.inbox === false) {
+      onIssueAction({
+        itemIds,
+        actionType: 'Reviewed',
+        shouldRemove: isForReviewQuery(query),
+      });
+      return;
+    }
+
+    if ('priority' in data && typeof data.priority === 'string') {
+      const priorityValues = parseIssuePrioritySearch(query);
+      const priority = data.priority.toLowerCase() as PriorityLevel;
+
+      onIssueAction({
+        itemIds,
+        actionType: 'Reprioritized',
+        shouldRemove: !priorityValues.has(priority),
+      });
+      return;
+    }
   };
 
-  render() {
-    if (
-      this.props.savedSearchLoading &&
-      !this.props.organization.features.includes('issue-stream-performance')
-    ) {
-      return this.renderLoading();
-    }
+  const onDelete = () => {
+    actionTakenRef.current = true;
+    fetchData(true);
+  };
 
-    const {
-      pageLinks,
-      queryCount,
-      queryCounts,
-      realtimeActive,
-      groupIds,
-      queryMaxCount,
-      issuesLoading,
-      error,
-    } = this.state;
-    const {organization, selection, router} = this.props;
-    const query = this.getQuery();
+  const paginationAnalyticsEvent = (direction: string) => {
+    trackAnalytics('issues_stream.paginate', {
+      organization,
+      direction,
+    });
+  };
 
-    const modifiedQueryCount = Math.max(queryCount, 0);
+  const onSavedSearchSelect = (newSavedSearch: SavedSearch) => {
+    trackAnalytics('organization_saved_search.selected', {
+      organization,
+      search_type: 'issues',
+      id: newSavedSearch.id ? parseInt(newSavedSearch.id, 10) : -1,
+      is_global: newSavedSearch.isGlobal,
+      query: newSavedSearch.query,
+      visibility: newSavedSearch.visibility,
+    });
+    setIssuesLoading(true);
+    setTimeout(() => {
+      transitionTo(undefined, newSavedSearch);
+    }, 0);
+  };
 
-    // TODO: these two might still be in use for reprocessing2
-    const showReprocessingTab = this.displayReprocessingTab();
-    const displayReprocessingActions = this.displayReprocessingLayout(
-      showReprocessingTab,
-      query
-    );
+  const modifiedQueryCount = Math.max(queryCount, 0);
 
-    const {numPreviousIssues, numIssuesOnPage} = this.getPageCounts();
+  // TODO: these two might still be in use for reprocessing2
+  const showReprocessingTab = !!queryCounts?.[Query.REPROCESSING]?.count;
+  const displayReprocessingActions = showReprocessingTab && query === Query.REPROCESSING;
 
-    return (
-      <NewTabContextProvider>
-        <Layout.Page>
-          {organization.features.includes('issue-stream-custom-views') ? (
-            <ErrorBoundary message={'Failed to load custom tabs'} mini>
-              <IssueViewsIssueListHeader
-                router={router}
-                selectedProjectIds={selection.projects}
-                realtimeActive={realtimeActive}
-                onRealtimeChange={this.onRealtimeChange}
-              />
-            </ErrorBoundary>
-          ) : (
-            <IssueListHeader
-              organization={organization}
-              query={query}
-              sort={this.getSort()}
-              queryCount={queryCount}
-              queryCounts={queryCounts}
-              realtimeActive={realtimeActive}
+  const {numPreviousIssues, numIssuesOnPage} = getPageCounts();
+
+  return (
+    <NewTabContextProvider>
+      <Layout.Page>
+        {organization.features.includes('issue-stream-custom-views') ? (
+          <ErrorBoundary message={'Failed to load custom tabs'} mini>
+            <IssueViewsIssueListHeader
               router={router}
-              displayReprocessingTab={showReprocessingTab}
               selectedProjectIds={selection.projects}
-              onRealtimeChange={this.onRealtimeChange}
+              realtimeActive={realtimeActive}
+              onRealtimeChange={onRealtimeChange}
             />
-          )}
+          </ErrorBoundary>
+        ) : (
+          <IssueListHeader
+            organization={organization}
+            query={query}
+            sort={sort}
+            queryCount={queryCount}
+            queryCounts={queryCounts}
+            realtimeActive={realtimeActive}
+            router={router}
+            displayReprocessingTab={showReprocessingTab}
+            selectedProjectIds={selection.projects}
+            onRealtimeChange={onRealtimeChange}
+          />
+        )}
 
-          <StyledBody>
-            <StyledMain>
-              <IssuesDataConsentBanner source="issues" />
-              <IssueListFilters
-                query={query}
-                sort={this.getSort()}
-                onSortChange={this.onSortChange}
-                onSearch={this.onSearch}
-              />
-              <IssueListTable
-                selection={selection}
-                query={query}
-                queryCount={modifiedQueryCount}
-                onSelectStatsPeriod={this.onSelectStatsPeriod}
-                onActionTaken={this.onActionTaken}
-                onDelete={this.onDelete}
-                statsPeriod={this.getGroupStatsPeriod()}
-                groupIds={groupIds}
-                allResultsVisible={this.allResultsVisible()}
-                displayReprocessingActions={displayReprocessingActions}
-                sort={this.getSort()}
-                onSortChange={this.onSortChange}
-                memberList={this.state.memberList}
-                selectedProjectIds={selection.projects}
-                issuesLoading={issuesLoading}
-                error={error}
-                refetchGroups={this.fetchData}
-                paginationCaption={
-                  !issuesLoading && modifiedQueryCount > 0
-                    ? tct('[start]-[end] of [total]', {
-                        start: numPreviousIssues + 1,
-                        end: numPreviousIssues + numIssuesOnPage,
-                        total: (
-                          <StyledQueryCount
-                            hideParens
-                            hideIfEmpty={false}
-                            count={modifiedQueryCount}
-                            max={queryMaxCount || 100}
-                          />
-                        ),
-                      })
-                    : null
-                }
-                pageLinks={pageLinks}
-                onCursor={this.onCursorChange}
-                paginationAnalyticsEvent={this.paginationAnalyticsEvent}
-                personalSavedSearches={this.props.savedSearches?.filter(
-                  search => search.visibility === 'owner'
-                )}
-                organizationSavedSearches={this.props.savedSearches?.filter(
-                  search => search.visibility === 'organization'
-                )}
-              />
-            </StyledMain>
-            <SavedIssueSearches
-              {...{organization, query}}
-              onSavedSearchSelect={this.onSavedSearchSelect}
-              sort={this.getSort()}
+        <StyledBody>
+          <StyledMain>
+            <IssuesDataConsentBanner source="issues" />
+            <IssueListFilters
+              query={query}
+              sort={sort}
+              onSortChange={onSortChange}
+              onSearch={onSearch}
             />
-          </StyledBody>
-        </Layout.Page>
-      </NewTabContextProvider>
-    );
-  }
+            <IssueListTable
+              selection={selection}
+              query={query}
+              queryCount={modifiedQueryCount}
+              onSelectStatsPeriod={onSelectStatsPeriod}
+              onActionTaken={onActionTaken}
+              onDelete={onDelete}
+              statsPeriod={getGroupStatsPeriod()}
+              groupIds={groupIds}
+              allResultsVisible={allResultsVisible()}
+              displayReprocessingActions={displayReprocessingActions}
+              sort={sort}
+              onSortChange={onSortChange}
+              memberList={memberList}
+              selectedProjectIds={selection.projects}
+              issuesLoading={issuesLoading}
+              error={error}
+              refetchGroups={fetchData}
+              paginationCaption={
+                !issuesLoading && modifiedQueryCount > 0
+                  ? tct('[start]-[end] of [total]', {
+                      start: numPreviousIssues + 1,
+                      end: numPreviousIssues + numIssuesOnPage,
+                      total: (
+                        <QueryCount
+                          hideParens
+                          hideIfEmpty={false}
+                          count={modifiedQueryCount}
+                          max={queryMaxCount || 100}
+                        />
+                      ),
+                    })
+                  : null
+              }
+              pageLinks={pageLinks}
+              onCursor={onCursorChange}
+              paginationAnalyticsEvent={paginationAnalyticsEvent}
+              personalSavedSearches={savedSearches?.filter(
+                search => search.visibility === 'owner'
+              )}
+              organizationSavedSearches={savedSearches?.filter(
+                search => search.visibility === 'organization'
+              )}
+            />
+          </StyledMain>
+          <SavedIssueSearches
+            {...{organization, query}}
+            onSavedSearchSelect={onSavedSearchSelect}
+            sort={sort}
+          />
+        </StyledBody>
+      </Layout.Page>
+    </NewTabContextProvider>
+  );
 }
 
-export default withRouteAnalytics(
-  withApi(
-    withPageFilters(
-      withSavedSearches(withOrganization(Sentry.withProfiler(IssueListOverview)))
-    )
-  )
-);
-
-export {IssueListOverview};
+export default Sentry.withProfiler(IssueListOverview);
 
 const StyledBody = styled('div')`
   background-color: ${p => p.theme.background};
@@ -1311,8 +1180,4 @@ const StyledMain = styled('section')`
   @media (min-width: ${p => p.theme.breakpoints.medium}) {
     padding: ${space(3)} ${space(4)};
   }
-`;
-
-const StyledQueryCount = styled(QueryCount)`
-  margin-left: 0;
 `;
